Catch and display errors during bootstrapping when using testdox printer
