+++ phpstan-tests.neon
@@ -16,6 +16,9 @@ parameters:
         # This access to undefined property is legit
         - '#Access to an undefined property SplObjectStorage::\$foo#'
 
+        # intentionally non existent function in tests/Regression/GitHub/3107/Issue3107Test.php
+        - '#Function does_not_exist not found#'
+
     excludes_analyse:
         # duplicated classname OneTest
         - tests/_files/phpunit-example-extension/tests/OneTest.php
